Handle XFS root filesystem present in F22
Fedora has switched to XFS as the root filesystem, even on armhfp
images :-/  XFS does not support shrinking and growing it is more
complex as well.

This commit adds limited support for XFS rootfs.  Resize is not
allowed (i.e. user-specified ROOTSIZE has no effect), but
the image can be created.

Fixes #6

Signed-off-by: Seth Jennings <sjennings@variantweb.net>