Adjust default sizes for Fedora 22
Fedora 22 minimal image rootfs is now over 1G and does not build
with the default options.

The original defaults were designed to build an image that would fit on
a 1GB card (if such a thing exists anymore).

The new default create an image that will fit comfortably on a
2GB card.

Fixes #7.

Signed-off-by: Seth Jennings <sjennings@variantweb.net>